Changed the custom status modal to class component Added recent custom statuses in the modal Fixed the styling of titles Fixed lint errors except one
